refactor:[sharding-ui-frontend]Add folder does not exist

+++ sharding-ui/sharding-ui-frontend/build/rimraf.js
@@ -388,13 +388,15 @@ function resolve(dir) {
 const nodeModulesPath = resolve('node_modules')
 
 fs.stat(nodeModulesPath, (err, stats)=> {
-  const stat = stats.isDirectory()
-  if(stat) {
-    const list = fs.readdirSync(nodeModulesPath)
-    if(!list.length) return
-    rimraf(nodeModulesPath, () => {
-      console.log('delete node_modules')
-    })
+  if(stats) {
+    const stat = stats.isDirectory()
+    if(stat) {
+      const list = fs.readdirSync(nodeModulesPath)
+      if(!list.length) return
+      rimraf(nodeModulesPath, () => {
+        console.log('delete node_modules')
+      })
+    }
   }
   return
 })
